% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbarplot.R
\name{varbarplot}
\alias{varbarplot}
\title{varbarplot}
\usage{
varbarplot(matrix, title, legend = FALSE)
}
\arguments{
\item{matrix}{a numeric matrix.}

\item{title}{a title for the barplot produced.}

\item{legend}{a logical indicating whether to plot a legend, default is FALSE.}
}
\value{
A barplot
}
\description{
Barplot of a numeric matrix from a VCF file.
}
\examples{
data(vslyzr_example)
chrom <- create.chromR('sc50', seq=dna, vcf=vcf, ann=gff)
chrom <- masker(chrom)
elems <- extract.gt(chrom, element="GQ", as.numeric=TRUE)
varbarplot(elems)
}
\seealso{
\code{\link{seq}}
}
