% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vslyzr_example.rdata}
\alias{vslyzr_example.rdata}
\title{Example data for vslyzr.}
\source{
\url{https://github.com/knausb/vcfR/blob/master/data/vcfR_example.rdata}
}
\description{
An example dataset is a subset of the pinfsc50 dataset.
This data consists of 3 samples and 5 variants.
}
\details{
\describe{
  \item dna DNAbin object
  \item gff gff format data.frame
  \item vcf vcfR object
}
}
\examples{
data(vslyzr_example)
\dontrun{
system.file("data", "vslyzr_example.rdata", package="vslyzr")
fPath <- system.file("data", "vslyzr_example.rdata", package="vslyzr")
}

}
